server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["adam-michael-bishop <adam.michael.bishop@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
motor = "^3.5.1"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
stripe = "^10.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
